ci: run the changelog rendering tests
The suite was not run in CI at all, so nothing checked the changelog
configuration. Most of these tests need a Git server and a crate registry,
which this workflow does not provide, but changelog generation is entirely
local and needs neither.

Move the rendering test into its own file so the workflow can select it by
target rather than by test name, and so a future rendering test is picked up
without also having to remember to edit the workflow.
